refactor(expeditions): extract view logic into child components
Moves creature selector filtering/display and expedition detail
rendering out of ExpeditionsView into dedicated CreatureSelector
and ExpeditionDetail components.
